Split sector name generation into separate file
